ICU-8/ Add UT for login attempt and create table
