+++ test/context.py
@@ -1,0 +1,7 @@
+import os
+import sys
+
+# add source files to path
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+from src import main
